Replace original files
