+++ client/src/app/pages/Login.js
@@ -1,65 +1,58 @@
 import React, {useState, useEffect} from 'react';
-import {useLocation, useNavigate} from 'react-router-dom';
-import {Stack} from '@mui/material';
+import {useNavigate} from 'react-router-dom';
 import {toast} from 'react-toastify';
-import '../stylesheets/Login.css';
-
-import verifyEmail from '../util/EmailVerification';
+import {InputContext} from '../components/ThemedInput';
+import Box from '@mui/material/Box';
+import Checkbox from '@mui/material/Checkbox';
+import FormControlLabel from '@mui/material/FormControlLabel';
+import Paper from '@mui/material/Paper';
+import ThemedButton from '../components/ThemedButton';
+import ThemedInput from '../components/ThemedInput';
+import LoginBanner from '../assets/LoginBanner.png';
 import useAuth from '../util/AuthContext';
+import '../stylesheets/LoginSignup.css';
+
+const PaperStyling = {
+  display: 'flex',
+  width: '1000px',
+  height: '600px',
+  borderRadius: '10px',
+  filter: 'drop-shadow(0px 15px 40px rgba(192, 225, 255, 0.1))',
+  color: '#3C4047',
+};
+
+const InputLabelStyling = {
+  '.MuiTypography-root': {
+    fontFamily: 'inherit',
+    fontSize: '0.8rem',
+    fontWeight: 'inherit',
+    color: '#8B95A5',
+  },
+  'marginLeft': '1em',
+};
 
 /**
- * creates login page
+ * Creates login page
  * @return {HTML} login page
  */
-export default function Login() {
+export default function TestLogin() {
   const navigate = useNavigate();
-  const location = useLocation();
-  const [signUp, setSignUp] = useState(false);
-  const [verifyEmailNotification, setVerifyEmailNotification] = useState(false);
-  const [createdProfileData, setCreatedProfileData] = useState(null);
-
   const {user, setUser, setLoggedIn, setUserProfile} = useAuth();
 
-  const [accountLoginCredentials, setAccountLoginCredentials] = useState({
-    useremail: '',
-    userpassword: '',
+  const [stepPage, setStepPage] = useState('login');
+  const [values, setValues] = useState({
+    'login': {
+      useremail: '',
+      userpassword: '',
+    },
+    'forgot1': {
+      useremail: '',
+    },
+    'forgot2': {
+      newpassword: '',
+      confirmpassword: '',
+    },
   });
-
-  const [newAccountCredentials, setnewAccountCredentials] = useState({
-    useremail: '',
-    userpassword: '',
-    active: 'false',
-  });
-
-  const handleChange = (e) => {
-    const {name, value} = e.target;
-    setAccountLoginCredentials({...accountLoginCredentials, [name]: value});
-  };
-
-  const handleChangeNewAccount = (e) => {
-    const {name, value} = e.target;
-    setnewAccountCredentials({...newAccountCredentials, [name]: value});
-  };
-
-  const handleEnter = (e) => {
-    if (e.key === 'Enter') {
-      login();
-    }
-  };
-
-  const handleEnterNewAccount = (e) => {
-    if (e.key === 'Enter') {
-      createUser();
-    }
-  };
-
-  const toggleSignup = () => {
-    setSignUp(!signUp);
-  };
-
-  useEffect(() => {
-    setSignUp(false);
-  }, [location.key, location.state]);
 
   useEffect(() => {
     if (user != null) {
@@ -67,59 +60,10 @@ export default function Login() {
     }
   }, [user]);
 
-
-  const handleResendClick = () => {
-    console.log('here');
-    console.log(createdProfileData);
-    verifyEmail(createdProfileData);
-  };
-
-  const createUser = () => {
-    fetch(`/api/userCreation`, {
-      method: 'POST',
-      body: JSON.stringify(newAccountCredentials),
-      headers: {
-        'Content-Type': 'application/json',
-      },
-    })
-        .then((res) => {
-          if (!res.ok) {
-            throw res;
-          }
-          return res.json();
-        })
-        .then((json) => {
-          fetch(`/api/profileCreation`, {
-            method: 'POST',
-            body: JSON.stringify({userid: json.userid}),
-            headers: {
-              'Content-Type': 'application/json',
-            },
-          });
-          toast.success('Account created', {
-            position: 'top-right',
-            autoClose: 5000,
-            hideProgressBar: false,
-            closeOnClick: true,
-            pauseOnHover: true,
-            draggable: true,
-            progress: undefined,
-          });
-          setCreatedProfileData(json);
-          verifyEmail(json);
-          setVerifyEmailNotification(true);
-          setnewAccountCredentials(
-              {useremail: '',
-                userpassword: '',
-                active: 'false'});
-          setSignUp(false);
-        });
-  };
-
   const login = () => {
     fetch(`/api/login`, {
       method: 'POST',
-      body: JSON.stringify(accountLoginCredentials),
+      body: JSON.stringify(values['login']),
       headers: {
         'Content-Type': 'application/json',
       },
@@ -149,10 +93,9 @@ export default function Login() {
         })
         .catch((err) => {
           console.log(err);
-          alert('Invalid Username or Password, Please Try Again');
+          alert('Invalid Username or Password. Please Try Again.');
         });
   };
-
 
   const getProfile = () => {
     fetch(`/api/getProfile/${user.userid}`)
@@ -173,106 +116,289 @@ export default function Login() {
         });
   };
 
-  return (
-    <div className="LoginPage">
-      {!signUp && !verifyEmailNotification &&
-      <div className="LoginPage__body">
-        <Stack>
-          <div className="LoginPage__topText">Log In</div>
-          <label className="LoginPage__label">
-              Email
-            <input
-              className="LoginPage__input"
-              name="useremail"
-              value={accountLoginCredentials.useremail}
-              onChange={handleChange}
-              onKeyDown={handleEnter}/>
-          </label>
-          <label className="LoginPage__label">
-              Password
-            <input
-              type="password"
-              className="LoginPage__input"
-              name="userpassword"
-              value={accountLoginCredentials.userpassword}
-              onChange={handleChange}
-              onKeyDown={handleEnter}
-            />
-          </label>
-          <button
-            className="LoginPage__submitButton"
-            onClick={login}
-          >
-            Submit
-          </button>
-          <button
-            className="LoginPage__submitButton"
-            onClick={toggleSignup}
-          >
-            Sign Up
-          </button>
-        </Stack>
-      </div>}
-      {signUp && !verifyEmailNotification &&
-      <div className="LoginPage__body">
-        <Stack>
-          <div className="LoginPage__topText">Sign Up</div>
-          <label className="LoginPage__label">
-          Email
-            <input
-              className="LoginPage__input"
-              name="useremail"
-              value={newAccountCredentials.useremail}
-              onChange={handleChangeNewAccount}
-              onKeyDown={handleEnterNewAccount}
-            />
-          </label>
-          <label className="LoginPage__label">
-          Password
-            <input
-              type="password"
-              className="LoginPage__input"
-              name="userpassword"
-              value={newAccountCredentials.userpassword}
-              onChange={handleChangeNewAccount}
-              onKeyDown={handleEnterNewAccount}
-            />
-          </label>
+  const handleNextPage = (step) => {
+    setStepPage(step);
+  };
 
-          <button
-            className="LoginPage__submitButton"
-            onClick={createUser}
-          >
-            Submit
-          </button>
-          <button
-            className="LoginPage__submitButton"
-            onClick={toggleSignup}
-          >
-            Log in
-          </button>
-        </Stack>
-      </div>}
-      {verifyEmailNotification && <div className='LoginPage__body'>
-        <Stack>
-          <div className='VerifyEmailNotification__topText'>
-            Verify Your Email
+  return (
+    <InputContext.Provider value={[values, setValues]}>
+      <Box className='page'>
+        <Paper className='card' elevation={0} sx={PaperStyling}>
+          <div className='card-banner flow-small padding-64'>
+            <p className='text-bold text-italic text-white'>Logo.</p>
+            <h3 className='text-xbold text-white'>Welcome back!</h3>
+            <img src={LoginBanner} />
           </div>
-          <p className='VerifyEmailNotification__message'>
-            {`We sent an email to ${createdProfileData.useremail}
-             to verify your email address and activate your account.
-             Please use the link in the email to verify your email address.
-             The link will expire in 48 hours.`}</p>
-          <p className='VerifyEmailNotification__message'>
-           If you did not recieve the email,
-            please click the link below to resend.
+          <Box
+            className='card-content padding-64'
+            component='form'
+            autoComplete='on'
+            noValidate
+          >
+            <LoginForm
+              active={stepPage === 'login'}
+              handleNextPage={(e) => handleNextPage(e)}
+              login={login}
+            />
+            <ForgotPasswordOne
+              active={stepPage === 'forgot1'}
+              handleNextPage={(e) => handleNextPage(e)}
+            />
+            <ForgotPasswordTwo
+              active={stepPage === 'forgot2'}
+              handleNextPage={(e) => handleNextPage(e)}
+            />
+            <ForgotPasswordThree
+              active={stepPage === 'forgot3'}
+              handleNextPage={(e) => handleNextPage(e)}
+            />
+          </Box>
+        </Paper>
+      </Box>
+    </InputContext.Provider>
+  );
+}
+
+/**
+ * Login form
+ * @return {JSX}
+ */
+function LoginForm({active, handleNextPage, login}) {
+  const navigate = useNavigate();
+
+  const handleNavigate = () => {
+    navigate('/signup');
+  };
+
+  return (
+    <div className='flow-large' style={{display: active ? null : 'none'}}>
+      <div className='grid-flow-large'>
+        <div>
+          <h2 className='text-normal'>Login</h2>
+          <p className='text-light text-warning'>
+            Required <span className='text-bold'>*</span>
           </p>
-          <div className='verifyEmailNotification__resendLink'
-            onClick={handleResendClick} style={{cursor: 'pointer'}}>
-              click here
-          </div>
-        </Stack>
-      </div>}
+        </div>
+        <p className='text-gray text-lineheight-24'>
+          Enter your email address and password
+          below to login to your account.
+        </p>
+      </div>
+      <div className='grid-flow-large'>
+        <div className='grid-flow-small'>
+          <p className='text-bold'>
+            Email <span className='text-bold text-warning'>*</span>
+          </p>
+          <ThemedInput
+            placeholder={'bobsmith@gmail.com'}
+            type={'text'}
+            index={'useremail'}
+            step={'login'}
+            fill={'email'}
+          />
+        </div>
+        <div className='grid-flow-small'>
+          <p className='text-bold'>
+            Password <span className='text-bold text-warning'>*</span>
+          </p>
+          <ThemedInput
+            placeholder={'Your password'}
+            type={'password'}
+            index={'userpassword'}
+            step={'login'}
+          />
+          <p
+            className='text-blue clickable'
+            onClick={(e) => handleNextPage('forgot1')}
+          >
+            Forgot your password?
+          </p>
+        </div>
+      </div>
+      <div className='grid-flow-small'>
+        <div>
+          <ThemedButton
+            color={'yellow'}
+            variant={'themed'}
+            type={'submit'}
+            onClick={(e) => {
+              e.preventDefault();
+              login();
+            }}
+          >
+            Login
+          </ThemedButton>
+          <FormControlLabel
+            label='Keep me logged in'
+            control={<Checkbox disableRipple />}
+            sx={InputLabelStyling}
+          />
+        </div>
+        <p className='text-light'>
+          Don&apos;t have an account?
+          <span
+            className='text-bold text-blue clickable'
+            onClick={handleNavigate}
+          >
+            &nbsp;Register here
+          </span>
+        </p>
+      </div>
+    </div>
+  );
+}
+
+/**
+ * Part one of changing password
+ * @return {JSX}
+ */
+function ForgotPasswordOne({active, handleNextPage}) {
+  return (
+    <div className='flow-large' style={{display: active ? null : 'none'}}>
+      <div className='grid-flow-large'>
+        <h2 className='text-normal'>Forgot your password?</h2>
+        <p className='text-gray text-lineheight-24'>
+          Don&apos;t worry, we can help you out! If you remember
+          your email address, you can quickly reset your password.
+          Input your email address and we&apos;ll send you a link to your
+          email that will allow you to reset your password.
+        </p>
+      </div>
+      <div className='grid-flow-small'>
+        <p className='text-bold'>Email</p>
+        <ThemedInput
+          placeholder={'bobsmith@gmail.com'}
+          type={'text'}
+          index={'useremail'}
+          step={'forgot1'}
+          fill={'email'}
+        />
+      </div>
+      <div className='grid-flow-small'>
+        <div className='flex-flow-large'>
+          <ThemedButton
+            color={'yellow'}
+            variant={'cancel'}
+            value={'login'}
+            onClick={(e) => handleNextPage(e.target.value)}
+          >
+            Back
+          </ThemedButton>
+          <ThemedButton
+            color={'yellow'}
+            variant={'themed'}
+            value={'forgot2'}
+            onClick={(e) => handleNextPage(e.target.value)}
+          >
+            Request password change
+          </ThemedButton>
+        </div>
+        <p className='text-light'>
+          Need help? Contact us at
+          <span className='text-bold text-blue'> tasselsupport@gmail.com</span>
+        </p>
+      </div>
+    </div>
+  );
+}
+
+/**
+ * Part two of changing password
+ * @return {JSX}
+ */
+function ForgotPasswordTwo({active, handleNextPage}) {
+  return (
+    <div className='flow-large' style={{display: active ? null : 'none'}}>
+      <div className='grid-flow-large'>
+        <h2 className='text-normal'>Change your password</h2>
+        <p className='text-gray text-lineheight-24'>
+          Enter your new password below.
+          We strongly advise you to store it safely.
+        </p>
+      </div>
+      <div className='grid-flow-large'>
+        <div className='grid-flow-small'>
+          <p className='text-bold'>New Password</p>
+          <ThemedInput
+            placeholder={'8+ Characters, 1 Capital Letter'}
+            type={'password'}
+            index={'newpassword'}
+            step={'forgot2'}
+          />
+        </div>
+        <div className='grid-flow-small'>
+          <p className='text-bold'>Confirm Password</p>
+          <ThemedInput
+            placeholder={'8+ Characters, 1 Capital Letter'}
+            type={'password'}
+            index={'newpassword'}
+            step={'forgot2'}
+          />
+        </div>
+      </div>
+      <div className='grid-flow-small'>
+        <div className='flex-flow-large'>
+          <ThemedButton
+            color={'yellow'}
+            variant={'cancel'}
+            value={'login'}
+            onClick={(e) => handleNextPage(e.target.value)}
+          >
+            Cancel
+          </ThemedButton>
+          <ThemedButton
+            color={'yellow'}
+            variant={'themed'}
+            value={'forgot3'}
+            onClick={(e) => handleNextPage(e.target.value)}
+          >
+            Change password
+          </ThemedButton>
+        </div>
+        <p className='text-light'>
+          Need help? Contact us at
+          <span className='text-bold text-blue'> tasselsupport@gmail.com</span>
+        </p>
+      </div>
+    </div>
+  );
+}
+
+/**
+ * Part three of changing password
+ * @return {JSX}
+ */
+function ForgotPasswordThree({active}) {
+  const navigate = useNavigate();
+
+  const handleNavigate = () => {
+    navigate('/login');
+  };
+
+  return (
+    <div className='flow-large' style={{display: active ? null : 'none'}}>
+      <div className='grid-flow-large text-center'>
+        <h2 className='text-normal'>Success!</h2>
+        <p className='text-gray text-lineheight-24'>
+          We have successfully changed your password.
+          Click the button below to login to your account.
+        </p>
+      </div>
+      <div className='grid-flow-small grid-center'>
+        <div className='flex-flow-small'>
+          <ThemedButton
+            color={'yellow'}
+            variant={'themed'}
+            onClick={handleNavigate}
+          >
+            Login
+          </ThemedButton>
+        </div>
+        <p className='text-light'>
+          Need help? Contact us at
+          <span className='text-bold text-blue'> tasselsupport@gmail.com</span>
+        </p>
+      </div>
     </div>
   );
 }
