--- client/src/app/stylesheets/Verify.css
@@ -1,10 +1,0 @@
-.verificationResponse {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.response {
-  display: flex;
-  justify-content: center;
-}
